Consistent smoke test package naming
