+++ test/smoke/testApps/Jdbc/src/smokeTest/java/com/microsoft/applicationinsights/smoketest/JdbcDisabledTest.java
@@ -1,0 +1,44 @@
+/*
+ * ApplicationInsights-Java
+ * Copyright (c) Microsoft Corporation
+ * All rights reserved.
+ *
+ * MIT License
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this
+ * software and associated documentation files (the ""Software""), to deal in the Software
+ * without restriction, including without limitation the rights to use, copy, modify, merge,
+ * publish, distribute, sublicense, and/or sell copies of the Software, and to permit
+ * persons to whom the Software is furnished to do so, subject to the following conditions:
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ * THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
+ * PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
+ * FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+ * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+ * DEALINGS IN THE SOFTWARE.
+ */
+
+package com.microsoft.applicationinsights.smoketest;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
+import org.junit.Test;
+
+@UseAgent("disabled_jdbc")
+public class JdbcDisabledTest extends AiWarSmokeTest {
+
+  @Test
+  @TargetUri("/hsqldbPreparedStatement")
+  public void hsqldbPreparedStatement() throws Exception {
+    Telemetry telemetry = getTelemetry(0);
+
+    assertEquals("GET /Jdbc/*", telemetry.rd.getName());
+    assertTrue(telemetry.rd.getSuccess());
+
+    // sleep a bit and make sure no jdbc dependencies are reported
+    Thread.sleep(5000);
+    assertEquals(0, mockedIngestion.getCountForType("RemoteDependencyData"));
+  }
+}
