+++ test/smoke/testApps/TelemetryFiltering/src/smokeTest/java/com/microsoft/applicationinsights/smoketest/TelemetryFiltering2SmokeTest.java
@@ -19,36 +19,17 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.smoketestapp;
+package com.microsoft.applicationinsights.smoketest;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-
-import com.microsoft.applicationinsights.smoketest.AiWarSmokeTest;
-import com.microsoft.applicationinsights.smoketest.DependencyContainer;
-import com.microsoft.applicationinsights.smoketest.TargetUri;
-import com.microsoft.applicationinsights.smoketest.UseAgent;
-import com.microsoft.applicationinsights.smoketest.WithDependencyContainers;
 import org.junit.Test;
 
-@UseAgent("disabled_mongo")
-@WithDependencyContainers(
-    @DependencyContainer(
-        value = "mongo:4",
-        portMapping = "27017",
-        hostnameEnvironmentVariable = "MONGO"))
-public class MongoDisabledTest extends AiWarSmokeTest {
+@UseAgent("telemetryfiltering2")
+public class TelemetryFiltering2SmokeTest extends AiWarSmokeTest {
 
   @Test
-  @TargetUri("/mongo")
-  public void mongo() throws Exception {
-    Telemetry telemetry = getTelemetry(0);
-
-    assertEquals("GET /MongoDB/*", telemetry.rd.getName());
-    assertTrue(telemetry.rd.getSuccess());
-
-    // sleep a bit and make sure no mongo dependencies are reported
-    Thread.sleep(5000);
-    assertEquals(0, mockedIngestion.getCountForType("RemoteDependencyData"));
+  @TargetUri(value = "/login", callCount = 100)
+  public void testSampling() throws Exception {
+    mockedIngestion.waitForItems("RequestData", 100);
+    mockedIngestion.waitForItems("RemoteDependencyData", 100);
   }
 }
